+++ python/src/aac/lang/hierarchy.py
@@ -4,7 +4,7 @@ import logging
 from typing import Optional
 
 from aac.lang.language_context import LanguageContext
-from aac.lang.definition_helpers import get_definition_by_name
+from aac.lang.definitions.collections import get_definition_by_name
 from aac.lang.definitions.definition import Definition
 from aac.lang.definitions.search import search_definition
