+++ python/src/aac/plugins/first_party/gen_plant_uml/gen_plant_uml_impl.py
@@ -4,7 +4,7 @@ import os
 from typing import Callable
 
 from aac.lang.active_context_lifecycle_manager import get_active_context
-from aac.lang.definition_helpers import get_definitions_by_root_key
+from aac.lang.definitions.collections import get_definitions_by_root_key
 from aac.lang.definitions.definition import Definition
 from aac.lang.definitions.search import search_definition
 from aac.plugins.plugin_execution import PluginExecutionResult, plugin_result
