+++ python/tests/plugins/validators/test_defined_references.py
@@ -1,5 +1,5 @@
 from aac.lang.active_context_lifecycle_manager import get_active_context
-from aac.lang.definition_helpers import get_definition_by_name, get_definitions_by_root_key
+from aac.lang.definitions.collections import get_definition_by_name, get_definitions_by_root_key
 from aac.plugins.validators import ValidatorFindings, ValidatorResult
 from aac.plugins.contributions.contribution_types import DefinitionValidationContribution
 from aac.plugins.validators.defined_references import _get_plugin_definitions, _get_plugin_validations, validate_references
