+++ python/tests/plugins/validators/test_root_keys.py
@@ -1,7 +1,7 @@
 import logging
 
 from aac.lang.active_context_lifecycle_manager import get_active_context
-from aac.lang.definition_helpers import get_definition_by_name, get_definitions_by_root_key
+from aac.lang.definitions.collections import get_definition_by_name, get_definitions_by_root_key
 from aac.plugins.contributions.contribution_types import DefinitionValidationContribution
 from aac.plugins.validators.root_keys import _get_plugin_definitions, _get_plugin_validations, validate_root_keys
