+++ python/tests/plugins/validators/test_validation_implementation_plugin.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from aac.lang.active_context_lifecycle_manager import get_active_context
-from aac.lang.definition_helpers import get_definitions_by_root_key
+from aac.lang.definitions.collections import get_definitions_by_root_key
 from aac.plugins.contributions.contribution_types import DefinitionValidationContribution
 from aac.plugins.validators.validator_implementation import _get_plugin_definitions, _get_plugin_validations, validate_validator_implementations
